bcc_detection/sample/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
Pillow>=8.0.0
matplotlib>=3.3.2
seaborn>=0.11.0
scikit-learn>=0.24.0
tqdm>=4.50.0
opencv-python>=4.5.1
scipy>=1.6.0
pandas>=1.1.3
tifffile>=2020.9.3 